Aborted attempt to bowlfishifh OpenBSD 6.4.
6.4 will never work because unveil, used in ifconfig and tcpdump,
requires a read-write root partition.  This is a flaw they have
addressed for later releases.

There's a bunch of other things in the way such as the varialbe paths
used in the object linker, but those can be worked around.

Saving what I had as it may be useful for future tries.